% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_chromosomal_names}
\alias{check_chromosomal_names}
\title{Check whether the two chromosome vectors are the same.}
\usage{
check_chromosomal_names(input_chrom, target_chrom)
}
\arguments{
\item{input_chrom}{The vector to be checked (as A).}

\item{target_chrom}{The target vector (as B).}
}
\value{
A == B: "identical"; A < B: "missing"; A > B: "extra"; as well as "other".
}
\description{
Check whether the two chromosome vectors are the same.
}
\examples{

check_chromosomal_names(input_chrom = LETTERS[1:5], target_chrom = LETTERS[1:26])
check_chromosomal_names(input_chrom = LETTERS[1:9], target_chrom = LETTERS[1:9])

}
