% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earlyORlate.R
\name{earlyORlate}
\alias{earlyORlate}
\title{Clonal and temporal dissection}
\usage{
earlyORlate(patient, complete.mutation.table, purity)
}
\arguments{
\item{patient}{patient describes the TCGA identifier}

\item{complete.mutation.table}{This is the mutation table in matrix format.}

\item{purity}{ASCAT purity estimate [range, 0.01,1]}
}
\value{
A data.frame.
}
\description{
This function implementents the cancer cell fraction analysis, as well as temporal dissection of mutations
}
\examples{

}
