\name{EstimateClonaliy-package}
\alias{EstimateClonality-package}
\alias{EStimateClonality}
\docType{package}
\title{
Temporal and clonal dissection of mutations
}
\description{
Uses read count and copy number information to temporally and clonally dissect SNVs.
}
\details{
\tabular{ll}{
Package: \tab EstimateClonality\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-12-23\cr
License: \tab What license is it under?\cr
}

}
\author{
Nicholas McGranahan

Maintainer: nicholas.mcgranahan@cancer.org.uk

}

\keyword{ package }
\seealso{
\code{\link{clonality.estimation},\link{earlyORlate}} 
}
\examples{
#see example.script for examples on how to implement main function
}
