% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GD.functions.R
\name{genome.doub.sig}
\alias{genome.doub.sig}
\title{Genome Doubling function}
\usage{
genome.doub.sig(sample, seg.mat.minor, seg.mat.copy, number.of.sim = 10000)
}
\arguments{
\item{sample}{TCGA sample identifier}

\item{seg.mat.minor}{segmented minor allele copy number matrix}

\item{seg.mat.copy}{segmented total copy number matrix}

\item{number.of.sim}{The number of simulations to assess genome doubling likelihood, defaults to 10,000}
}
\value{

}
\description{
The following function estimates the probability that a genome doubling has
occured at some point during the evolutionary history of a tumour.
}
\details{
According to the code of clonality.estimation(), "val" means "cn".
}
\examples{
genome.doub.sig(sample, seg.mat.minor, seg.mat.copy, number.of.sim = 10000)

}
